Working on boxes more. Added -h/--noheader to new.js.

+++ src/haunted/control.cpp
@@ -1,11 +1,11 @@
 #include "haunted/control.h"
 
 namespace haunted {
-	ssize_t control::max_children() const {
+	int control::max_children() const {
 		return -1;
 	}
 
-	size_t control::child_count() const {
+	int control::child_count() const {
 		return children.size();
 	}
 }
